+++ Proyecto_final/tercera_entrega/src/controller/web/products.controller.js
@@ -9,50 +9,60 @@ async function getProduct(req, res) {
 
 async function postProduct(req, res) {
     const { title, price, thumbnail, desc, stock } = req.body
-    if (!title|| !price || !thumbnail || !desc || !stock) return res.send('Completar todos los campos')
+    try {
+        if (!title|| !price || !thumbnail || !desc || !stock) throw new Error('Completar todos los campos')
+        let added = await productsDao.add({ title, price, thumbnail, desc, stock })
+    
+        res.status(201).send(`Agregado: ${added}`)
 
-    let added = await productsDao.add({ title, price, thumbnail, desc, stock })
-
-    res.status(201).send(`Agregado: ${added}`)
+    } catch (error){
+        res.json({error: error.message})
+    }
 }
 
 async function putProduct(req, res){
     const { id } = req.params
     const {title, price, thumbnail, desc, stock} = req.body
-    if (id === -1) res.status(404).send('No se encontró el producto')
+    try {
+        let productToUpdate = await productsDao.getById(id)
+        if (!productToUpdate) throw new Error('cannot find product')
 
-    const updated = await productsDao.update(id, {title, price, thumbnail, desc, stock})
+        const updated = await productsDao.update(id, {title, price, thumbnail, desc, stock})
+        res.status(200).json(updated)
 
-    res.status(200).json(updated)
+    } catch (error){
+        res.json({error: error.message})
+    }
+
 }
 
 async function delProduct(req, res){
     const {id} = req.params
+    try {
+        let productToDelete = await productsDao.getById(id)
+        if(!productToDelete) throw new Error('cannot find product')
+    
+        const deleted = await productsDao.delete(id)
+    
+        res.send(deleted)
 
-    let idMax = await productsDao.getAll()
-
-    if (id > idMax.length) {
-        res.status(404).send({error: 'el producto no existe'})
+    } catch (error){
+        res.json({error: error.message})
     }
-
-    const deleted = await productsDao.delete(id)
-
-    res.send(deleted)
 }
 
 async function getById(req, res){
-    
     const { id } = req.params
+    try {
+        let product = await productsDao.getById(id)
+        if(!product) throw new Error('cannot find product')
+    
+        res.status(200).send(product)
 
-    let idMax = await productsDao.getAll()
-
-    if (id > idMax.length) {
-        res.status(404).send({error: 'el producto no existe'})
+    } catch (error){
+        res.json({error: error.message})
     }
 
-    let product = await productsDao.getById(id)
-
-    res.status(200).send(product)
 }
 
 // -------------------------------
